Only show force edit button if payment status is 2

+++ src/app/page-restricted-area/page-manage-major-events/page-list-subscriptions/page-manage-subscription/page-manage-subscription.page.ts
@@ -67,6 +67,9 @@ export class PageManageSubscriptionPage implements OnInit {
         .get()
         .subscribe((doc) => {
           const data = doc.data();
+          if (data.payment.status !== 2) {
+            return;
+          }
           data.subscribedToEvents.forEach((event) => {
             this.afs.doc<EventItem>(`events/${event}/subscriptions/${this.subscriptionID}`).delete();
             this.afs.doc<EventItem>(`events/${event}`).update({
